// src/models/RemoteOrderHistoryDetail.ts
export type RemoteOrderHistoryDetail = {
    serialNumber: string;
    ticketIdx: string;
    gameName: string;
    drawIdx: string;
    drawEndDate: string;
    linePrice: number;
    currencyType: string;
    totalLineCount: string;
    paymentMethod: string;
    totalPrice: string;
    orderStatus: number;
    createdAt: string;
};